refactor: restructure codebase for PyPi distribution as pygitgo
