--- installer/linux-installer/install.sh
@@ -1,182 +1,0 @@
-#!/bin/sh
-# GitGo Installer for Linux/Termux
-# Cross-platform installation script
-
-set -e  # Exit on error
-
-RED='\033[0;31m'
-GREEN='\033[0;32m'
-YELLOW='\033[1;33m'
-BLUE='\033[0;34m'
-NC='\033[0m' # No Color
-
-print_info() {
-    printf "${BLUE}[INFO]${NC} %s\n" "$1"
-}
-
-print_success() {
-    printf "${GREEN}[SUCCESS]${NC} %s\n" "$1"
-}
-
-print_warning() {
-    printf "${YELLOW}[WARNING]${NC} %s\n" "$1"
-}
-
-print_error() {
-    printf "${RED}[ERROR]${NC} %s\n" "$1"
-}
-
-print_header() {
-    echo ""
-    echo "========================================"
-    echo "   GitGo Automated Installer v1.0"
-    echo "========================================"
-    echo ""
-}
-
-is_termux() {
-    if [ -n "$PREFIX" ] && [ -d "/data/data/com.termux" ]; then
-        return 0
-    fi
-    return 1
-}
-
-get_python_cmd() {
-    if command -v python3 >/dev/null 2>&1; then
-        echo "python3"
-    elif command -v python >/dev/null 2>&1; then
-        echo "python"
-    else
-        echo ""
-    fi
-}
-
-main() {
-    print_header
-    
-    if is_termux; then
-        PLATFORM="Termux"
-        INSTALL_DIR="${PREFIX}/share/gitgo"
-        BIN_DIR="${PREFIX}/bin"
-    else
-        PLATFORM="Linux"
-        INSTALL_DIR="$HOME/.local/share/gitgo"
-        BIN_DIR="$HOME/.local/bin"
-    fi
-    
-    print_info "Platform detected: $PLATFORM"
-    
-    PYTHON_CMD=$(get_python_cmd)
-    if [ -z "$PYTHON_CMD" ]; then
-        print_error "Python not found!"
-        print_error "Please install Python 3 and try again."
-        exit 1
-    fi
-    
-    print_info "Python detected: $($PYTHON_CMD --version)"
-    
-    # Verify source files exist
-    SCRIPT_DIR="$(cd "$(dirname "$0")" && pwd)"
-    SRC_DIR="$SCRIPT_DIR/../../src"
-    
-    if [ ! -f "$SRC_DIR/gitgo.py" ]; then
-        print_error "gitgo.py not found in src directory!"
-        print_error "Make sure you're running this from the GitGo root directory."
-        exit 1
-    fi
-    
-    print_info "Installing to: $INSTALL_DIR"
-    
-    if [ -d "$INSTALL_DIR" ]; then
-        print_warning "Installation directory already exists. Removing old installation..."
-        rm -rf "$INSTALL_DIR"
-    fi
-    
-    mkdir -p "$INSTALL_DIR"
-    print_success "Created installation directory"
-    
-    print_info "Copying files..."
-    cp -r "$SRC_DIR/"* "$INSTALL_DIR/"
-    
-    if [ $? -ne 0 ]; then
-        print_error "Failed to copy files!"
-        exit 1
-    fi
-    
-    print_success "Files copied successfully"
-
-    if [ ! -d "$BIN_DIR" ]; then
-        mkdir -p "$BIN_DIR"
-        print_success "Created bin directory: $BIN_DIR"
-    fi
-    
-    WRAPPER_PATH="$BIN_DIR/gitgo"
-    GITGO_PY_PATH="$INSTALL_DIR/gitgo.py"
-    
-    print_info "Creating wrapper script..."
-    
-    cat > "$WRAPPER_PATH" << EOF
-#!/bin/sh
-# GitGo CLI Wrapper
-# Auto-generated by GitGo installer
-
-exec $PYTHON_CMD "$GITGO_PY_PATH" "\$@"
-EOF
-    
-    # Make wrapper executable
-    chmod +x "$WRAPPER_PATH"
-    print_success "Wrapper script created: $WRAPPER_PATH"
-    
-    case ":$PATH:" in
-        *":$BIN_DIR:"*)
-            print_success "Directory is already in PATH!"
-            ;;
-        *)
-            print_warning "Directory is NOT in PATH!"
-            print_warning "Add this directory to your PATH:"
-            print_success "    $BIN_DIR"
-            echo ""
-            print_info "Add this line to your shell config file (~/.bashrc or ~/.zshrc):"
-            printf "${GREEN}    export PATH=\"\$PATH:%s\"${NC}\n" "$BIN_DIR"
-            echo ""
-            print_info "Then run: source ~/.bashrc (or ~/.zshrc)"
-            echo ""
-            ;;
-    esac
-    
-    echo ""
-    print_info "Testing installation..."
-    
-    export PATH="$PATH:$BIN_DIR"
-    
-    if command -v gitgo >/dev/null 2>&1; then
-        if gitgo -r >/dev/null 2>&1; then
-            print_success "GitGo is working correctly!"
-            echo ""
-            echo "========================================"
-            echo "   Installation Complete!"
-            echo "========================================"
-            echo ""
-            echo "You can now use 'gitgo' from anywhere:"
-            echo "  gitgo -r                    (check status)"
-            echo "  gitgo push main \"message\"   (push to branch)"
-            echo "  gitgo push -n branch \"msg\"  (create new branch)"
-            echo ""
-        else
-            print_warning "GitGo command found but may not be working properly."
-            print_warning "Try running: gitgo -r"
-        fi
-    else
-        print_warning "GitGo command not found in PATH."
-        print_warning "You may need to update your PATH as shown above."
-        print_warning "After updating PATH, restart your terminal and run: gitgo -r"
-    fi
-    
-    if ! case ":$PATH:" in *":$BIN_DIR:"*) true;; *) false;; esac; then
-        echo ""
-        print_warning "NOTE: You need to add $BIN_DIR to your PATH"
-        print_warning "and restart your terminal before using gitgo."
-    fi
-}
-
-main
