+++ src/pygitgo/commands/git_operations.py
@@ -1,6 +1,6 @@
-from auth.ssh_utils import convert_https_to_ssh, is_ssh_url, check_connection
-from utils.executor import run_command
-from utils.colors import info, success, warning, error
+from pygitgo.auth.ssh_utils import convert_https_to_ssh, is_ssh_url, check_connection
+from pygitgo.utils.executor import run_command
+from pygitgo.utils.colors import info, success, warning, error
 import subprocess
 import sys
 import os
@@ -16,7 +16,6 @@ def get_current_branch():
     return branch
 
 def is_branch_exist(branch):
-    # This one check if the branch existed from both remote and local branches
     return bool(run_command(["git", "branch", "-r", "--list", f"*/{branch}"])) or bool(run_command(["git", "branch", "--list", branch]))
 
 
@@ -45,10 +44,8 @@ def git_init():
         warning("Already a git repository! Skipping init...")
         return True
     
-    # Use -b main to set default branch to 'main' (Git >= 2.28)
     result = run_command(["git", "init", "-b", DEFAULT_MAIN_BRANCH], allow_fail=True, loading_msg="Initializing git repository...")
     if isinstance(result, subprocess.CalledProcessError):
-        # Fallback for older Git versions that don't support -b
         run_command(["git", "init"], loading_msg="Initializing git repository...")
         run_command(["git", "checkout", "-b", DEFAULT_MAIN_BRANCH], allow_fail=True)
 
@@ -57,7 +54,6 @@ def git_init():
 
 
 def add_remote_origin(repo_url):
-    # clean the URL - remove quotes if present, dili na kailangan og quotes
     clean_url = repo_url.strip('"\'')
     
     existing_remote = run_command(["git", "remote", "get-url", "origin"], allow_fail=True)
@@ -71,7 +67,6 @@ def add_remote_origin(repo_url):
 
 
 def confirm_remote_link():
-    # test connection to remote - make sure it's not cap
     test_result = run_command(["git", "ls-remote", "origin"], allow_fail=True, loading_msg="Testing connection to remote...")
     
     if isinstance(test_result, subprocess.CalledProcessError):
@@ -84,14 +79,11 @@ def confirm_remote_link():
 
 
 def create_main_branch():
-    # check current branch - switch to main if needed
     current_branch = run_command(["git", "branch", "--show-current"], allow_fail=True)
     
     if isinstance(current_branch, subprocess.CalledProcessError) or not current_branch.strip():
-        # no commits yet, set default branch to main
         run_command(["git", "checkout", "-b", "main"], loading_msg="Setting default branch to 'main'...")
     elif current_branch.strip() != "main":
-        # rename current branch to main
         run_command(["git", "branch", "-m", "main"], loading_msg=f"Renaming branch '{current_branch.strip()}' to 'main'...")
     else:
         success("Already on 'main' branch.")
@@ -126,13 +118,11 @@ def check_and_sync_branch(branch):
 
 
 def git_push(branch):
-    # Check if remote is using HTTPS and SSH is available
     remote_url = run_command(["git", "remote", "get-url", "origin"], allow_fail=True)
     
     if not isinstance(remote_url, subprocess.CalledProcessError) and remote_url:
         remote_url = remote_url.strip()
         
-        # If using HTTPS and SSH is set up, convert to SSH
         if not is_ssh_url(remote_url) and check_connection():
             ssh_url = convert_https_to_ssh(remote_url)
             if ssh_url:
