+++ src/pygitgo/utils/executor.py
@@ -1,8 +1,7 @@
-from utils.colors import error, info, success, warning
-from yaspin import yaspin
 import subprocess
 import sys
 import os
+from pygitgo.utils.colors import error, info, success, warning
 
 
 def run_command(command, allow_fail=False, return_complete=False, loading_msg=None):
@@ -12,12 +11,11 @@ def run_command(command, allow_fail=False, return_complete=False, loading_msg=No
     :param command: list of command + args
     :param allow_fail: if True, do not exit on error
     :param return_complete: if True, return subprocess.CompletedProcess instead of stdout
+    :param loading_msg: if provided, show a yaspin spinner with this message
     """
+    from yaspin import yaspin
 
-    spinner = yaspin(
-        text=loading_msg,
-        color="cyan"
-    ) if loading_msg else None
+    spinner = yaspin(text=loading_msg, color="cyan") if loading_msg else None
 
     if spinner:
         spinner.start()
@@ -31,23 +29,20 @@ def run_command(command, allow_fail=False, return_complete=False, loading_msg=No
         )
 
         if spinner:
-            spinner.ok('✔')
+            spinner.ok("✔")
 
         return result if return_complete else result.stdout.strip()
     except subprocess.CalledProcessError as e:
+        if spinner:
+            spinner.fail("✖")
+
         stderr = e.stderr.strip() if e.stderr else ""
         
-        if spinner:
-            spinner.fail('✖')
-
-        # Check for dubious ownership error (common in Termux/shared storage)
         if "detected dubious ownership" in stderr:
             error(f"\nSECURITY ALERT: {stderr}")
             warning("\nThis is a known Git security feature in shared environments (like Termux).")
             info("GitGo can fix this for you by adding this directory to your safe list.")
             
-            # Extract the path from the error message if possible
-            # Error looks like: fatal: detected dubious ownership in repository at '/path/to/repo'
             import re
             path_match = re.search(r"repository at '(.+)'", stderr)
             repo_path = path_match.group(1) if path_match else os.getcwd()
@@ -62,7 +57,6 @@ def run_command(command, allow_fail=False, return_complete=False, loading_msg=No
                     subprocess.run(fix_command, check=True)
                     success("Success! Directory added to safe list.")
                     info("Retrying your original command...\n")
-                    # Recursive call to retry the original command
                     return run_command(command, allow_fail, return_complete, loading_msg=loading_msg)
                 except Exception as fix_err:
                     error(f"Failed to apply fix: {fix_err}")
@@ -70,7 +64,7 @@ def run_command(command, allow_fail=False, return_complete=False, loading_msg=No
                 warning("\nFix declined. Operations in this directory will continue to fail.")
 
         if allow_fail:
-            return e  # let caller decide
+            return e
         
         error(f"\nCommand Failed: {' '.join(command)}")
         error(f"Error output:\n{stderr or 'No error details available'}\n")
